site_kubectl/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pydantic
pyyaml
httpx
